+++ lib/danger/commands/plugin_abstract.rb
@@ -1,0 +1,5 @@
+module Danger
+  class PluginAbstract < Runner
+    require 'danger/commands/plugin_lint'
+  end
+end