+++ lib/danger/plugin_support/plugin_parser.rb
@@ -5,22 +5,23 @@ module Danger
     attr_accessor :registry
 
     def initialize(path)
-      @path = path
+      raise "Path cannot be empty" if path.empty?
+      @path = File.expand_path(path)
     end
 
     def parse
       # could this go in a singleton-y place instead?
       # like class initialize?
       YARD::Tags::Library.define_tag('tags', :tags)
-
       files = ["lib/danger/plugin_support/plugin.rb", @path]
+
+      # This turns on YARD debugging
+      # $DEBUG = true
       self.registry = YARD::Registry.load(files, true)
     end
 
     def classes_in_file
-      self.registry.all
-          .select { |thing| thing.type == :class }
-          .select { |klass| klass.file == @path }
+      registry.all(:class).select { |klass| klass.file == @path }
     end
 
     def plugins_from_classes(classes)
@@ -28,6 +29,7 @@ module Danger
     end
 
     def to_dict(classes)
+
       d_meth = lambda do |meth|
         return nil if meth.nil?
         {
@@ -50,20 +52,25 @@ module Danger
       end
 
       classes.map do |klass|
+        # Adds the class being parsed into the ruby runtime
+        puts klass.file
+        require klass.file
+        real_klass = Danger.const_get klass.name
+        attribute_meths = klass.attributes[:instance].values.map { |v| v.values }.flatten
+
         {
-        name: klass.name.to_s,
-        body_md: klass.docstring,
-        example_code: klass.tags.select { |t| t.tag_name == "example" }.map(&:text).compact,
-        attributes: klass.attributes[:instance].map do |pair|
-          {
-          pair.first => d_attr.call(pair.last)
+          name: klass.name.to_s,
+          body_md: klass.docstring,
+          instance_name: real_klass.instance_name,
+          example_code: klass.tags.select { |t| t.tag_name == "example" }.map(&:text).compact,
+          attributes: klass.attributes[:instance].map do |pair|
+            { pair.first => d_attr.call(pair.last) }
+          end,
+          methods: (klass.meths - klass.inherited_meths - attribute_meths ).select { |m| m.visibility == :public }.map { |m| d_meth.call(m) },
+          tags: klass.tags.select { |t| t.tag_name == "tags" }.map(&:name).compact,
+          see: klass.tags.select { |t| t.tag_name == "see" }.map(&:name).map(&:split).flatten.compact,
+          file: klass.file.gsub(File.expand_path("."), "")
         }
-        end,
-        methods: (klass.meths - klass.inherited_meths).select { |m| m.visibility == :public }.map { |m| d_meth.call(m) },
-        tags: klass.tags.select { |t| t.tag_name == "tags" }.map(&:name).compact,
-        see: klass.tags.select { |t| t.tag_name == "see" }.map(&:name).map(&:split).flatten.compact,
-        file: klass.file
-      }
       end
     end
   end