+++ spec/lib/danger/commands/plugin_lint_spec.rb
@@ -1,0 +1,10 @@
+require 'danger/commands/plugin_lint'
+
+module Danger
+  describe Danger::PluginLint do
+    it "runs the command" do
+      # allow(STDOUT).to receive(:puts) # this disables puts
+      Danger::PluginLint.run(["spec/fixtures/plugins/example_fully_documented.rb"])
+    end
+  end
+end