+++ lib/danger/commands/plugin_abstract.rb
@@ -1,5 +1,10 @@
 module Danger
   class PluginAbstract < Runner
     require 'danger/commands/plugin_lint'
+    require 'danger/commands/new_plugin'
+
+    self.command = 'plugin'
+
+    self.abstract_command = true
   end
 end