+++ lib/danger/danger_core/plugins/dangerfile_import_plugin.rb
@@ -1,9 +1,35 @@
 require 'danger/plugin_support/plugin'
 
 module Danger
+  # One way to support internal plugins is via `plugin.import` this gives you
+  # the chance to quickly iterate without the need for building rubygems. As such,
+  # it does not have the stringent rules around documentation expected of a public plugin.
+  # It's worth noting, that you can also have plugins inside `./danger_plugins` and they
+  # will be automatically imported into your Dangerfile at launch.
+  #
+  # @example Import a plugin available over HTTP
+  #
+  #          device_grid = "https://raw.githubusercontent.com/fastlane/fastlane/master/danger-device_grid/lib/device_grid/plugin.rb"
+  #          plugin.import device_grid
+  #
+  # @example Import from a local file reference
+  #
+  #          plugin.import "danger/plugins/watch_plugin.rb"
+  #
+  # @example Import all files inside a folder
+  #
+  #          plugin.import "danger/plugins/*.rb"
+  #
+  # @see  danger/danger
+  # @tags core, plugins
+
   class DangerfileImportPlugin < Plugin
+    def self.instance_name
+      "plugin"
+    end
+
     # @!group Plugins
-    # Download a local or remote plugin and use it locally
+    # Download a local or remote plugin and use it inside the Dangerfile.
     #
     # @param    [String] path
     #           a local path or a https URL to the Ruby file to import
@@ -19,8 +45,10 @@ module Danger
       end
     end
 
+    private
+
     # @!group Plugins
-    # Download a remote plugin and use it locally
+    # Download a remote plugin and use it locally.
     #
     # @param    [String] url
     #           https URL to the Ruby file to use
@@ -43,14 +71,15 @@ module Danger
     end
 
     # @!group Plugins
-    # Import one or more local plugins
+    # Import one or more local plugins.
     #
     # @param    [String] path
     #           The path to the file to import
     #           Can also be a pattern (./**/*plugin.rb)
     def import_local(path)
       Dir[path].each do |file|
-        require File.expand_path(file) # without the expand_path it would fail if the path doesn't start with ./
+        # Without the expand_path it would fail if the path doesn't start with ./
+        require File.expand_path(file)
         refresh_plugins
       end
     end