+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - a list of gems
   - no arguments, which will parse `lib/**/**/*` to lint your local plugins
 * Moved new plugin to `danger plugin new` - orta
+* Added `api` to the DSL, which is a shortcut to the active `Octokit::Client` - orta
+* Renamed `branch_for_merge` to `branch_for_base` and also added `branch_for_head` - orta
+* Initial work on namespacing existing plugins - orta
 * Notify the user to add the new GitHub account as collaborator to Close Source project
 
 ## 0.8.3