+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Initial work on namespacing existing plugins - orta
 * Notify the user to add the new GitHub account as collaborator to Close Source project
 * Fixes a problem running `danger local` due to a missing dependency on yard - ashfurrow
+* Improvements for CircleCI CI detection - orta
+
 
 ## 0.8.3